Add FontAwesome icons, refactor AboutMe component, and implement ButtonLinks; added basic stlying
