+++ src/App.jsx
@@ -1,12 +1,13 @@
 import "./index.css"
-import "font-awesome/css/font-awesome.min.css"
-import {AboutMe} from "./Sections/Main/About/AboutMe.jsx"
+import "./Sections/Main/About/AboutMe.jsx"
+import { AboutMe } from "./Sections/Main/About/AboutMe.jsx"
 
 
 export const App = () => {
+  console.log("App.jsx");
   return (
     <section>
-     <AboutMe/>
+      <AboutMe/>
     </section>
-  )
+  );
 }
