+++ src/Components/ButtonLinks/ButtonLinks.css
@@ -1,0 +1,13 @@
+ul {
+display: flex;
+flex-direction: row;
+justify-content: start;
+ list-style-type: none;
+}
+li {
+  margin: 0 .5rem;}
+a {
+  color: var(--font-color-light);
+  font-size: 1.5rem;
+  font-weight: 600;
+}
