+++ src/Components/Portrait/Portrait.css
@@ -1,0 +1,4 @@
+
+img{
+width: 20rem;
+height: auto;}
