+++ src/Components/Portrait/Portrait.jsx
@@ -1,12 +1,14 @@
-import photo from "../assets/img/Jhicks-PP.png";
+import photo from "../../assets/img/Jhicks-PP.png";
+import "./Portrait.css";
 
 
 const Portrait = () => {
 return (
     <div className="portrait">
         <img src={photo} alt="Portrait of Jonny Hicks" />
-<span id="circleBehind"></span>
-</div>
+        <span id="circleBehind">
+        </span>
+    </div>
 )}
 export default Portrait;
 // export default Portrait;
