+++ src/Sections/Main/About/AboutMe.css
@@ -1,0 +1,8 @@
+.about-me {
+display: flex;
+justify-content: start;
+flex-direction:row;
+}
+.about-text{
+width: 50vw;
+text-wrap: wrap;}
