+++ src/Sections/Main/About/AboutMe.jsx
@@ -1,15 +1,24 @@
-import Portrait from "../../../componants/Portrait/Portrait"; 
+import Portrait from "../../../Components/Portrait/Portrait";
+import "./AboutMe.css";
+import { ButtonLinks } from "../../../Components/ButtonLinks/ButtonLinks";
 
 
-const AboutMe = () => { 
-return (
-<div className="about-me">
-<div className="about-text">
-<h1>Jonny Hicks</h1>
-<h2> I Am Web Devolper</h2>
-<p>Some Text</p>
-</div>
-<div className="portait"><Portrait/></div>
-</div>)}
+export const AboutMe = () => {
+  return (
+	<div className="about-me">
 
-export default AboutMe;
+	  <div className="about-text">
+		  <h1>Jonny Hicks</h1>
+		  <h2>I Am a Web Developer</h2>
+		  <p>Some Text</p>
+      <ButtonLinks/>
+	  </div>
+
+	  <div className="portrait" >
+		  <Portrait/>
+	  </div>
+
+	</div>
+  );
+};
+
